+++ applications/fishing-map/src/features/timebar/timebar.hooks.ts
@@ -5,7 +5,7 @@ import { debounce } from 'lodash'
 import { ApiEvent } from '@globalfishingwatch/api-types/dist'
 import { TimebarVisualisations } from 'types'
 import { selectTimebarVisualisation } from 'features/app/app.selectors'
-import { useLocationConnect, useLoginRedirect } from 'routes/routes.hook'
+import { CALLBACK_URL_KEY, useLocationConnect } from 'routes/routes.hook'
 import {
   selectActiveActivityDataviews,
   selectActiveTrackDataviews,
@@ -28,7 +28,8 @@ export const TimeRangeAtom = atom<Range | null>({
   default: null,
   effects_UNSTABLE: [
     ({ trigger, setSelf, onSet }) => {
-      const { redirectUrl } = useLoginRedirect()
+      const redirectUrl =
+        typeof window !== undefined ? window.localStorage.getItem(CALLBACK_URL_KEY) : null
       const urlTimeRange = selectUrlTimeRange(store.getState() as RootState)
       const dispatch = useDispatch()
 
@@ -38,13 +39,17 @@ export const TimeRangeAtom = atom<Range | null>({
             ...urlTimeRange,
           })
         } else if (redirectUrl) {
-          // Workaround to get start and end date from redirect url as the
-          // location reducer isn't ready until initialDispatch
-          const url = new URL(redirectUrl)
-          const start = url.searchParams.get('start')
-          const end = url.searchParams.get('end')
-          if (start && end) {
-            setSelf({ start, end })
+          try {
+            // Workaround to get start and end date from redirect url as the
+            // location reducer isn't ready until initialDispatch
+            const url = new URL(JSON.parse(redirectUrl))
+            const start = url.searchParams.get('start')
+            const end = url.searchParams.get('end')
+            if (start && end) {
+              setSelf({ start, end })
+            }
+          } catch (e) {
+            console.warn(e)
           }
         }
       }